cleanup & multithreading preparation

+++ CPPNanoStringTests/nano64File1Tests.cpp
@@ -5,11 +5,11 @@
 TEST(nano64BytesFile1, creation)
 {
 	{
-		const auto ns{ nnstr::NanoString::make<64, "orange">() };
-		EXPECT_EQ(*ns, 2);
+		const auto ns{ nnstr::NanoString::make<size_t(64), "orange">() };
+		EXPECT_EQ(*ns, static_cast<size_t>(2));
 	}
 	{
-		const auto ns{ nnstr::NanoString::make<64, "potato">() };
-		EXPECT_EQ(*ns, 4);
+		const auto ns{ nnstr::NanoString::make<size_t(64), "potato">() };
+		EXPECT_EQ(*ns, static_cast<size_t>(4));
 	}
 }
